requirements.txt
streamlit
opencv-python-headless
numpy
pandas
pdf2image
cairosvg
svgwrite
Pillow